refactor: fix 18 trivial cross-context import violations
Replace deep internal import paths with imports from context public APIs
where the symbol was already exported. Ratchet exception budgets down.

Fully cleaned (0 violations remaining):
- SessionLifecycleManager.py - all 5 agent_sessions imports consolidated

Partially cleaned (budget ratcheted):
- artifacts.py: 7 -> 4
- executions/commands.py: 3 -> 2
- sessions.py: 5 -> 3
- triggers/commands.py: 11 -> 7
- workflows/commands.py: 10 -> 8
- seeding.py: 4 -> 3

Also adds TYPE_CHECKING static imports to orchestration/domain/__init__.py
so pyright can resolve aggregate types through the public API path.
